fix: ParentHome hasAlreadyBeenFetched 함수 react hook으로 통합
